[ew_substrate_fixed] rename source file and its test to `parse_fixed_point`
