+++ packages/ew_substrate_fixed/lib/src/parse_fixed_point.dart
